services/licensing/pyproject.toml
[project]
name = "actorhub-licensing-service"
version = "1.0.0"
description = "Content licensing microservice for ActorHub.ai"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
